BE-7: Added validation

+++ tests/test_user.py
@@ -2,6 +2,14 @@ import pytest
 
 
 @pytest.mark.asyncio
+async def test_login(test_client):
+    response = await test_client.post("/token", data={"username": "test@test.com", "password": "testpassword"})
+    assert response.status_code == 200
+    token = response.json()["access_token"]
+    assert token is not None
+    return token
+
+@pytest.mark.asyncio
 async def test_get_users(test_client):
     response = await test_client.get("/users/")
     assert response.status_code == 200
@@ -15,7 +23,7 @@ async def test_create_user(test_client):
         "user_firstname": "Test",
         "user_lastname": "User",
         "user_city": "User",
-        "user_phone": "User",
+        "user_phone": "+44 7911 123456",
         "user_avatar": "User",
     })
     assert response.status_code == 200
@@ -44,7 +52,8 @@ async def test_get_non_existing_user(test_client):
 
 @pytest.mark.asyncio
 async def test_update_user(test_client):
-    response = await test_client.put("/users/1", json={
+    token = await test_login(test_client)
+    response = await test_client.put("/users/1", headers={"Authorization": f"Bearer {token}"}, json={
           "user_firstname": "Updated",
           "user_lastname": "User",
           "user_city": "User",
@@ -57,6 +66,15 @@ async def test_update_user(test_client):
 
 
 @pytest.mark.asyncio
+async def test_delete_wrong_user(test_client):
+    token = await test_login(test_client)
+    response = await test_client.delete("/users/2", headers={"Authorization": f"Bearer {token}"},)
+    assert response.status_code == 403
+
+
+@pytest.mark.asyncio
 async def test_delete_user(test_client):
-    response = await test_client.delete("/users/1")
+    token = await test_login(test_client)
+    response = await test_client.delete("/users/1", headers={"Authorization": f"Bearer {token}"},)
     assert response.status_code == 200
+
